requirements.txt
Flask>=2.3.0
Flask-SQLAlchemy>=3.0.0
Flask-Login>=0.6.3
Flask-WTF>=1.1.0
python-dotenv>=1.0.0
Werkzeug>=2.3.0
requests>=2.31.0
beautifulsoup4>=4.12.0
